move folders to projects, packages for db added on solution level
