+++ Common/Enum/MessageType.cs
@@ -1,0 +1,9 @@
+namespace JourneyBot.Common.Enums
+{
+    public enum MessageType : byte
+    {
+        SimpleMessage = 0,
+
+        Interaction = 1
+    }
+}
