+++ JourneyBot.Logic/Interfaces/IMessageRenderer.cs
@@ -1,0 +1,11 @@
+using JourneyBot.Common.Enums;
+using JourneyBot.Datamodel.Models;
+
+namespace JourneyBot.Logic.Interfaces
+{
+    interface IMessageRenderer
+    {
+        SimpleMessageResponse RenderMessage(MessageType messageType, SimpleMessageOptions options);
+        InteractionResponse RenderMessage(MessageType messageType, InteractionOptions options);
+    }
+}
