+++ JourneyBot.Logic/JourneyBot.Logic.csproj
@@ -1,0 +1,14 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net7.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Common\JourneyBot.Common.csproj" />
+    <ProjectReference Include="..\JourneyBot.Datamodel\JourneyBot.Datamodel.csproj" />
+  </ItemGroup>
+
+</Project>
