--- interfaces/IMessageRenderer.cs
@@ -1,11 +1,0 @@
-
-interface IMessageRenderer
-{
-  SimpleMessageResponse RenderMessage(MessageType messageType, SimpleMessageOptions options);
-  InteractionResponse RenderMessage(MessageType messageType, InteractionOptions options);
-}
-
-
-
-
-
